requirements.txt
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pymongo==4.10.1
pyTelegramBotAPI==4.26.0
python-decouple==3.8
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
telebot==0.0.5
urllib3==2.3.0
